* Updated sources after Chrome version update; * Using std::unique_ptr instead of raw pointers; there's no STLDeleteElements anymore; * Added the android_aar script to generate .aar packages from the new GN build; * Adding the iOS device Name on the User-Agent header; * BUG Fix: the Authentication and Proxy-Authentication headers don't copy the parameters when cloned; * Adding +sip.instance only on REGISTER requests, as well as Supported header.

+++ sippet/transport/chrome/chrome_datagram_reader.h
@@ -26,8 +26,8 @@ class ChromeDatagramReader
   int DoIORead(const net::CompletionCallback& callback) override;
   char *data() override;
   size_t max_size() override;
-  int BytesRemaining() const override;
-  void DidConsume(int bytes) override;
+  size_t BytesRemaining() const override;
+  void DidConsume(size_t bytes) override;
 
   void OnReceiveDataComplete(int result);
   void ReceivedData(size_t bytes);
