+++ sippet/transport/chrome/chrome_stream_reader.h
@@ -27,8 +27,8 @@ class ChromeStreamReader
   int DoIORead(const net::CompletionCallback& callback) override;
   char *data() override;
   size_t max_size() override;
-  int BytesRemaining() const override;
-  void DidConsume(int bytes) override;
+  size_t BytesRemaining() const override;
+  void DidConsume(size_t bytes) override;
 
   void ReceiveDataComplete(int result);
   void DoCallback(int result);
